+++ tuna/miopen/driver/base.py
@@ -44,37 +44,31 @@ from tuna.driver import DriverBase
 LOGGER = setup_logger('driver_base')
 
 
+# pylint: disable=too-many-instance-attributes
 class MIOpenDriver(DriverBase):
   """Represents db tables based on ConfigType"""
-  tensor_attr: List[str] = [column.name for column in inspect(TensorTable).c]
-  tensor_id_map: Dict[str, int] = {}
+  tensor_attr: List[str]
 
-  def __init__(self,
-               line: str = str(),
-               db_obj: ConvolutionConfig = None) -> None:
-    if line:
-      if not self.construct_driver(line):
-        raise ValueError(f"Error creating Driver from line: '{line}'")
-    elif db_obj:
-      if not self.construct_driver_from_db(db_obj):
-        raise ValueError(
-            f"Error creating Driver from db obj: '{db_obj.to_dict()}'")
-    else:
-      raise ValueError(
-          "Error creating Driver. MIOpen Driver cmd line or db_obj required")
+  def __init__(self, line: str = "", db_obj: ConvolutionConfig = None):
+    super().__init__()
+    self.tensor_attr = [column.name for column in inspect(TensorTable).c]
 
+  @abstractmethod
   def parse_fdb_key(self, line: str):
     """Overloaded method.Defined in conv&bn driver child class"""
     raise NotImplementedError("Not implemented")
 
+  @abstractmethod
   def parse_row(self, db_obj: ConvolutionConfig):
     """Overloaded method.Defined in conv&bn driver child class"""
     raise NotImplementedError("Not implemented")
 
+  @abstractmethod
   def set_cmd(self, data_type: str):
     """Overloaded method.Defined in conv&bn driver child class"""
     raise NotImplementedError("Not implemented")
 
+  @abstractmethod
   def config_set_defaults(self):
     """Overloaded method.Defined in conv&bn driver child class"""
     raise NotImplementedError("Not implemented")
